% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkdata.R
\name{checkdata}
\alias{checkdata}
\title{Checks whether the given argument meets the minimum requirements for PrePRe functions}
\usage{
checkdata(pddt)
}
\arguments{
\item{pddt}{the data.table or data.frame that will be tested}
}
\value{
\code{NULL} if \code{pddt} does not meet the minimum requirements, or the \code{data.table} sorted by the \code{Time} column.
}
\description{
\code{checkdata} checks whether the given argument is a \code{data.table}, and if not converts the
argument to a \code{data.table}, checks whether the \code{data.table} contains the required columns
(i.e., both \code{Dil} and \code{Time}), and ensures that the \code{data.table} is sorted correctly
as the PrePRe functions assume that all dilation samples are stored sequentially.
}
\details{
Note that this function is not exported, it main purpose is to be used internally by PrePRe functions.
}

