% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotBlink.R
\name{plotBlink}
\alias{plotBlink}
\title{Plots a blink plus surrounding data and an indication of the cutoff thresholds}
\usage{
plotBlink(pddt, blink, expand, cutoff = NA)
}
\arguments{
\item{pddt}{A pddt data frame}

\item{blink}{Which blink to display}

\item{expand}{How much data (in ms) to plot before and after the blink}

\item{cutoff}{Where to plot the visual indicators for the cutoff (if NA, no indicators will be plotted)}
}
\description{
Creates a plot showing a blink (sequential number defined by argument \code{blink}) including
some data before and after the blink (\code{expand}, defined in miliseconds) and plots (if
specified) two vertical lines indicating potential cutoff criteria.
}
\author{
Hedderik van Rijn
}

