% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare eyetrack data table for analysis.}
\usage{
prepare(dvData, eye = "autoselect")
}
\arguments{
\item{dvData}{a data.frame or data.table as generated by SR-Research's DataViewer.}

\item{eye}{a character string indicating whether the \code{left} or \code{right} eye
information has to be retained, or whether this has to be \code{autodetect}ed.}
}
\value{
Returns a \code{data.table} with, at least, the following columns:
\itemize{
\item{\code{Time}} time points at which he samples were recorded (the original column \code{time})
\item{\code{X}} x-positions of the selected eye
\item{\code{Y}} y-positions of the selected eye
\item{\code{Dil}} recorded dilation values of the selected eye
}
The \code{data.table} is sorted by \code{Time}.
}
\description{
This function prepares an eyetrack data table for subsequent analyses. It modifies the data.table
in the following ways:
\itemize{
\item{It converts the column names to CamelCase}
\item{It converts the input to a data.table, if necessary}
\item{It checks whether the data.table contains all necessary columns}
\item{It retains the columns of either the right, or the left eye, either based on the argument
\code{eye}, or by assessing which column in the data.table does not contain any information.}
\item{It renames the coordinate columns to \code{X} and \code{Y}, the dilation column to
\code{Dil}, and the \code{Timestamp} column to \code{Time}.}
}
}
\author{
Hedderik van Rijn
}

