% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepreInfo.R
\name{prepreInfo}
\alias{prepreInfo}
\title{Prints the version information for the prepre package}
\usage{
prepreInfo(print = TRUE)
}
\arguments{
\item{print}{if \code{TRUE}, print version information to the console}
}
\value{
\code{prepreInfo} returns a character string containing the version and
 revision number of the package.
}
\description{
Prints the version, revision, and date information for the prepre package
}
\details{
This function prints the version and revision information for the prepre
package.
}
\author{
Based on BFInfo by Richard D. Morey, modified for prepre by Hedderik van Rijn
}
\keyword{misc}

