% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showBlinks.R
\name{showBlinks}
\alias{showBlinks}
\title{Start the Shiny Blink Inspector.}
\usage{
showBlinks(pddt)
}
\description{
Starts a Shiny app to allow for the inspection of blinks. Main purpose is to see how much data
around the blinks need to be rejected.
}
\author{
Hedderik van Rijn
}

