% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toCamelCase.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Converts a vector of text strings with elements deliminated by a separator to CamelCase}
\usage{
toCamelCase(x, sep = "_")
}
\arguments{
\item{x}{vector of character strings to convert}

\item{sep}{separator character between words in input vector}
}
\value{
vector of converted character strings
}
\description{
Converts a vector of text strings consisting of multiple word combinations separated by
underscores (or an other given character) to CamelCase, indicating that every word starts
with a capital, and all other letters are converted to lowercase.
}
\author{
Hedderik van Rijn
}

