% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numBlinks.R
\name{numBlinks}
\alias{numBlinks}
\title{Returns the number of blinks in the given data table.}
\usage{
numBlinks(pddt)
}
\description{
Returns the number of blinks in the given data table.
}
\author{
Hedderik van Rijn
}

