\name{prepre-package}
\alias{prepre-package}
\alias{prepre}
\docType{package}
\title{
PREprocessing of Pupillary REsponse data
}
\description{
Functions for preprocessing of pupillary data, including blink detection, blink interpolation, baselining, etc. 
}
\details{
\tabular{ll}{
Package: \tab prepre\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-06\cr
License: \tab GPL \cr
}

As the continuous sampling of the pupillary response results in large data sets, this package relies on the data.table package. All functions assume that pupillary data is stored in a data.table that contains the following columns:




}
\author{
Hedderik van Rijn
Jacolien van Rij

Maintainer: Please sent all comments to <vr-vr@van-rijn.org>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
\code{\link[<data.table>:<data.table>-package]{<data.table>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
